+++ slsDetectorSoftware/slsDetector/slsDetectorActions.h
@@ -198,7 +198,7 @@ class slsDetectorActions : public virtual slsDetectorBase
 
 
   virtual int setThresholdEnergy(int, int im=-1, detectorSettings isettings=GET_SETTINGS)=0;
-  virtual int setChannel(long long, int ich=-1, int ichip=-1, int imod=-1)=0;
+  virtual int setChannel(int64_t, int ich=-1, int ichip=-1, int imod=-1)=0;
 
 
   int setStartIndex(int i=-1){if (i>=0) {startIndex=i; lastIndex=startIndex; nowIndex=startIndex;};return startIndex;};
